pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "monitors-mcp"
version = "0.1.0"
description = "Claude Monitor System - MCP Server for distributed monitoring"
authors = [
    {name = "9bikes Team", email = "team@9bikes.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.12.3",
    "pydantic>=2.10.6",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "sqlalchemy>=2.0.23",
    "pymysql>=1.1.0",
    "redis>=5.0.1",
    "influxdb-client>=1.41.0",
    "cryptography>=41.0.7",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.2",
    "httpx>=0.25.2",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.1",
]

[project.scripts]
monitors-mcp = "monitors_mcp.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 140
target-version = ["py311"]

[tool.ruff]
target-version = "py311"
line-length = 140

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
